requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_pytorch_select=0.2=gpu_0
_tflow_select=2.3.0=mkl
absl-py=0.11.0=pyhd3eb1b0_1
antlr4-python3-runtime=4.8=pypi_0
astor=0.8.1=py36_0
backports=1.0=py_2
backports.functools_lru_cache=1.6.4=pyhd8ed1ab_0
blas=1.0=mkl
boto3=1.16.39=pyhd3eb1b0_0
botocore=1.19.39=pyhd3eb1b0_0
brotlipy=0.7.0=py36h27cfd23_1003
c-ares=1.17.1=h27cfd23_0
ca-certificates=2020.12.5=ha878542_0
certifi=2020.12.5=py36h5fab9bb_1
cffi=1.14.4=py36h261ae71_0
chardet=3.0.4=pypi_0
colorama=0.4.4=pypi_0
contextlib2=0.6.0.post1=pypi_0
cryptography=3.3.1=py36h3c74f83_0
cudatoolkit=10.0.130=0
cudnn=7.6.5=cuda10.0_0
cupy=8.3.0=py36h91b26fc_0
cycler=0.10.0=py36_0
cymem=1.31.2=py36_0
cython=0.29.21=pypi_0
cytoolz=0.8.2=py36_0
dataclasses=0.8=pypi_0
dateparser=1.0.0=pypi_0
dbus=1.13.18=hb2f20db_0
decorator=5.0.7=pyhd3eb1b0_0
dgl-cuda10.0=0.5.3=py36_0
dgl-cuda10.1=0.6.1=py36_0
dill=0.3.3=pypi_0
docopt=0.6.2=pypi_0
docutils=0.14=py36_0
en-core-web-sm=1.2.0=pypi_0
expat=2.2.10=he6710b0_2
fairseq=1.0.0a0+ac11107=dev_0
faiss-gpu=1.6.5=pypi_0
fastrlock=0.6=py36h2531618_0
filelock=3.0.12=pypi_0
fontconfig=2.13.0=h9420a91_0
freetype=2.10.4=h5ab3b9f_0
ftfy=4.4.3=py36_0
future=0.18.2=pypi_0
fuzzywuzzy=0.18.0=pypi_0
gast=0.4.0=py_0
gensim=3.8.3=pypi_0
glib=2.66.1=h92f7085_0
grpcio=1.31.0=py36hf8bcb03_0
gst-plugins-base=1.14.0=hbbd80ab_1
gstreamer=1.14.0=hb31296c_0
h5py=2.10.0=py36hd6299e0_1
hdf5=1.10.6=hb1b8bf9_0
html5lib=1.1=pyh9f0ad1d_0
hydra-core=1.0.4=pypi_0
hyperopt=0.1.2=pypi_0
icu=58.2=he6710b0_3
idna=2.7=pypi_0
imageio=2.4.1=py36_0
importlib-metadata=2.0.0=py_1
importlib-resources=3.3.0=pypi_0
intel-openmp=2020.2=254
jmespath=0.10.0=py_0
joblib=0.17.0=py_0
jpeg=9b=h024ee3a_2
json-tricks=3.15.5=pypi_0
keras-applications=1.0.8=py_1
keras-preprocessing=1.1.0=py_1
kiwisolver=1.3.0=py36h2531618_0
lcms2=2.11=h396b838_0
ld_impl_linux-64=2.33.1=h53a641e_7
libedit=3.1.20191231=h14c3975_1
libffi=3.3=he6710b0_2
libgcc-ng=9.1.0=hdf63c60_0
libgfortran=3.0.0=1
libgfortran-ng=7.3.0=hdf63c60_0
libpng=1.6.37=hbc83047_0
libprotobuf=3.13.0.1=hd408876_0
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.1.0=h2733197_1
libuuid=1.0.3=h1bed415_2
libxcb=1.14=h7b6447c_0
libxml2=2.9.10=hb55368b_3
lz4-c=1.9.2=heb0550a_3
markdown=3.3.3=py36h06a4308_0
matplotlib=2.2.2=py36hb69df0a_2
mkl=2020.2=256
mkl-service=2.3.0=py36he8ac12f_0
mkl_fft=1.2.0=py36h23d657b_0
mkl_random=1.1.1=py36h0573a6f_0
multiprocess=0.70.11.1=pypi_0
murmurhash=0.26.4=py36_0
nccl=2.9.8.1=h86b6200_0
ncurses=6.2=he6710b0_1
netifaces=0.10.9=pypi_0
networkx=2.2=py36_1
ninja=1.10.2=py36hff7bd54_0
nni=2.2=pypi_0
numpy=1.19.2=py36h54aff64_0
numpy-base=1.19.2=py36hfa32c7d_0
olefile=0.46=py36_0
omegaconf=2.0.5=pypi_0
openblas=0.2.19=0
openssl=1.1.1k=h27cfd23_0
packaging=20.8=pypi_0
pandas=1.1.3=py36he6710b0_0
pathlib=1.0.1=py36h5fab9bb_4
pathos=0.2.7=pypi_0
patsy=0.5.1=py36_0
pcre=8.44=he6710b0_0
pillow=5.3.0=pypi_0
pip=20.3.1=pypi_0
pipreqs=0.4.11=pypi_0
plac=0.9.6=py36_0
portalocker=2.0.0=pypi_0
pox=0.2.9=pypi_0
ppft=1.6.6.3=pypi_0
preshed=1.0.0=py36_0
prettytable=2.1.0=pypi_0
protobuf=3.13.0.1=py36he6710b0_1
psutil=5.8.0=pypi_0
pycparser=2.20=py_2
pymongo=3.11.4=pypi_0
pyopenssl=20.0.1=pyhd3eb1b0_1
pyparsing=2.4.7=py_0
pyqt=5.9.2=py36h05f1152_2
pysocks=1.7.1=py36h06a4308_0
python=3.6.12=hcff3b4d_2
python-dateutil=2.8.1=py_0
python_abi=3.6=1_cp36m
pythonwebhdfs=0.2.3=pypi_0
pytorch=1.4.0=cuda100py36h75a84eb_0
pytz=2020.4=pyhd3eb1b0_0
pyyaml=5.3.1=pypi_0
pyzmq=17.1.2=pypi_0
qt=5.9.7=h5867ecd_1
readline=8.0=h7b6447c_0
regex=2017.4.5=py36_0
requests=2.19.1=pypi_0
responses=0.13.3=pypi_0
ruamel-yaml=0.17.4=pypi_0
ruamel-yaml-clib=0.2.2=pypi_0
s3transfer=0.3.3=pyhd3eb1b0_3
sacrebleu=1.4.14=pypi_0
sacremoses=0.0.43=pypi_0
schema=0.7.4=pypi_0
scikit-learn=0.24.2=pypi_0
scipy=1.1.0=pypi_0
seaborn=0.9.0=py36_0
sentencepiece=0.1.94=pypi_0
setuptools=51.0.0=py36h06a4308_2
simplejson=3.17.2=pypi_0
sip=4.19.8=py36hf484d3e_0
six=1.11.0=pypi_0
smart-open=4.0.1=pypi_0
spacy=1.8.2=py36_0
sqlite=3.33.0=h62c20be_0
statsmodels=0.12.1=py36h27cfd23_0
swig=3.0.12=h38cdd7d_3
tensorboard=1.12.2=py36he6710b0_0
tensorboardx=2.1=py_0
tensorflow=1.12.0=mkl_py36h69b6ba0_0
tensorflow-base=1.12.0=mkl_py36h3c3e929_0
termcolor=1.1.0=py36_1
thinc=6.5.2=py36_0
threadpoolctl=2.1.0=pyh5ca1d4c_0
tk=8.6.10=hbc83047_0
tokenizers=0.7.0=pypi_0
toolz=0.11.1=py_0
torchfile=0.1.0=pypi_0
torchnet=0.0.4=pypi_0
torchvision=0.2.1=py36_0
tornado=5.1.1=pypi_0
tqdm=4.54.1=pypi_0
transformers=2.9.1=pypi_0
typing-extensions=3.7.4.3=pypi_0
tzlocal=2.1=pypi_0
ujson=1.35=py36h201607c_1003
urllib3=1.23=pypi_0
visdom=0.1.8.5=pypi_0
wcwidth=0.2.5=pyh9f0ad1d_2
webencodings=0.5.1=py_1
websocket-client=0.53.0=pypi_0
websockets=9.0.2=pypi_0
werkzeug=1.0.1=py_0
wget=3.2=pypi_0
wheel=0.36.1=pyhd3eb1b0_0
word2vec=0.9.4=py36h6323ea4_0
wrapt=1.12.1=py36h1d69622_2
xz=5.2.5=h7b6447c_0
yarg=0.1.9=pypi_0
zipp=3.4.0=pyhd3eb1b0_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.5=h9ceee32_0
